+++ packages/ente-ui/src/routes/Entries/CreateEntry.tsx
@@ -51,6 +51,7 @@ import {
 } from "ente-types";
 import { DateInput } from "../../elements/DateInput";
 import { createTranslation } from "ente-ui/src/helpers/createTranslation";
+import { Maybe } from "monet";
 
 const lang = createTranslation({
   en: {
@@ -105,10 +106,10 @@ interface CreateEntryOwnProps {
 }
 
 interface CreateEntryStateProps {
-  isParent: boolean;
-  children: UserN[];
+  isParent: Maybe<boolean>;
+  children: Maybe<UserN[]>;
   teachers: UserN[];
-  getUser(id: string): UserN;
+  getUser(id: string): Maybe<UserN>;
 }
 const mapStateToProps: MapStateToPropsParam<
   CreateEntryStateProps,
@@ -149,7 +150,7 @@ interface State {
 class CreateEntry extends React.Component<CreateEntryProps, State> {
   state: State = {
     student:
-      this.props.children.length > 0
+      this.props.children.some().length > 0
         ? this.props.children[0].get("id")
         : undefined,
     isRange: false,
@@ -262,14 +263,11 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
   minDate: Date = twoWeeksBefore(new Date());
 
   render() {
-    const { isParent } = this.props;
+    const { isParent, fullScreen, onClose, show, children } = this.props;
+    const { isRange, student, forSchool, date, dateEnd, slots } = this.state;
 
     return (
-      <Dialog
-        fullScreen={this.props.fullScreen}
-        onClose={this.props.onClose}
-        open={this.props.show}
-      >
+      <Dialog fullScreen={fullScreen} onClose={onClose} open={show}>
         <DialogTitle>{lang.newEntry}</DialogTitle>
         <DialogContent>
           <Grid container direction="column" spacing={40}>
@@ -279,7 +277,7 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
                   <FormControlLabel
                     control={
                       <Checkbox
-                        checked={this.state.isRange}
+                        checked={isRange}
                         onChange={this.handleChangeIsRange}
                       />
                     }
@@ -290,7 +288,7 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
                   <FormControlLabel
                     control={
                       <Checkbox
-                        checked={this.state.forSchool}
+                        checked={forSchool}
                         onChange={this.handleChangeForSchool}
                       />
                     }
@@ -298,17 +296,17 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
                   />
                 </Grid>
               </Grid>
-              {isParent && (
+              {isParent.some() && (
                 <Grid item>
                   <TextField
                     fullWidth
                     select
                     label={lang.child}
-                    value={this.state.student}
+                    value={student}
                     onChange={this.handleChangeStudent}
                     helperText={lang.selectChild}
                   >
-                    {this.props.children.map(child => (
+                    {children.some().map(child => (
                       <MenuItem key={child.get("id")} value={child.get("id")}>
                         {child.get("displayname")}
                       </MenuItem>
@@ -324,17 +322,17 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
                       isValid={this.dateValid}
                       onChange={this.handleChangeBeginDate}
                       minDate={this.minDate}
-                      value={this.state.date}
+                      value={date}
                     />
                   </Grid>
-                  {this.state.isRange && (
+                  {isRange && (
                     <Grid item xs={6}>
                       <DateInput
                         label={lang.toLabel}
                         isValid={this.dateEndValid}
                         onChange={this.handleChangeDateEnd}
-                        minDate={nextDay(this.state.date)}
-                        value={this.state.dateEnd}
+                        minDate={nextDay(date)}
+                        value={dateEnd}
                         minDateMessage={lang.dateMustBeBiggerThanFrom}
                       />
                     </Grid>
@@ -346,7 +344,7 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
               <Typography variant="h6">{lang.titles.slots}</Typography>
               <Typography variant="caption">{lang.addSlotsCaption}</Typography>
               <MUIList>
-                {this.state.slots.map((slot, index) => (
+                {slots.map((slot, index) => (
                   <SlotListItem
                     key={index}
                     slot={slot}
@@ -356,24 +354,24 @@ class CreateEntry extends React.Component<CreateEntryProps, State> {
               </MUIList>
               <SlotEntry
                 onAdd={slot => this.handleAddSlot(slot)}
-                multiDay={this.state.isRange}
+                multiDay={isRange}
                 datePickerConfig={{
-                  min: this.state.date,
-                  max: this.state.dateEnd
+                  min: date,
+                  max: dateEnd
                 }}
-                date={this.state.date}
+                date={date}
               />
             </Grid>
           </Grid>
         </DialogContent>
         <DialogActions>
-          <Button onClick={this.props.onClose} color="secondary">
+          <Button onClick={onClose} color="secondary">
             {lang.cancel}
           </Button>
           <Button
             onClick={() => {
               this.handleSubmit();
-              this.props.onClose();
+              onClose();
             }}
             disabled={!this.inputValid()}
             color="primary"