+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Changed: API now takes entry date for slots as well, if it is a single-day entry
+* Changed: Redux selectors now return maybe types in some cases
+
 * Fixed: Import window now accepts all mime types and instead filters by file ending (must end on ".csv")
 * Changed: Password Reset screen now redirects to login screen
 * Changed: Login screen now shows feedback while pending requests